--- src/junkyard/2glabstraction.jl
@@ -1,33 +1,0 @@
-
-
-#Came from GLWindow/screen.jl
-
-#this is definetely useful, namechange to something that isn't a callback, more like 
-# opengl_catch_error or something
-
-"""
-Callback which can be used to catch OpenGL errors.
-"""
-function openglerrorcallback(
-        source::GLenum, typ::GLenum,
-        id::GLuint, severity::GLenum,
-        length::GLsizei, message::Ptr{GLchar},
-        userParam::Ptr{Void}
-    )
-    errormessage = """
-         ________________________________________________________________
-        |
-        | OpenGL Error!
-        | source: $(GLENUM(source).name) :: type: $(GLENUM(typ).name)
-        |  $(String(message, length))
-        |________________________________________________________________
-    """
-    output = typ == GL_DEBUG_TYPE_ERROR ? error : info
-    output(errormessage)
-    nothing
-end
-
-global const _openglerrorcallback = cfunction(
-    openglerrorcallback, Void,
-    (GLenum, GLenum,GLuint, GLenum, GLsizei, Ptr{GLchar}, Ptr{Void})
-)
