--- src/junkyard/defaults.jl
@@ -1,6 +1,0 @@
-
-function draw_fullscreen(vao_id)
-    glBindVertexArray(vao_id)
-    glDrawArrays(GL_TRIANGLES, 0, 3)
-    glBindVertexArray(0)
-end
