--- src/junkyard/postprocessing.jl
@@ -1,62 +1,0 @@
-
-struct PostprocessPrerender
-end
-function (sp::PostprocessPrerender)()
-    glDepthMask(GL_TRUE)
-    glDisable(GL_DEPTH_TEST)
-    glDisable(GL_BLEND)
-    glDisable(GL_STENCIL_TEST)
-    glStencilMask(0xff)
-    glDisable(GL_CULL_FACE)
-    nothing
-end
-
-loadshader(name) = joinpath(dirname(@__FILE__), name)
-
-
-rcpframe(x) = 1f0./Vec2f0(x[1], x[2])
-
-"""
-Creates a postprocessing render object.
-This will transfer the pixels from the color texture of the Framebuffer
-to the screen and while at it, it can do some postprocessing (not doing it right now):
-E.g fxaa anti aliasing, color correction etc.
-"""
-function postprocess(color, color_luma, framebuffer_size)
-    shader1 = LazyShader(
-        loadshader("fullscreen.vert"),
-        loadshader("postprocess.frag")
-    )
-    data1 = Dict{Symbol, Any}(
-        :color_texture => color
-    )
-    pass1 = RenderObject(data1, shader1, PostprocessPrerender(), nothing)
-    pass1.postrenderfunction = () -> draw_fullscreen(pass1.vertexarray.id)
-    shader2 = LazyShader(
-        loadshader("fullscreen.vert"),
-        loadshader("fxaa.frag")
-    )
-    data2 = Dict{Symbol, Any}(
-        :color_texture => color_luma,
-        :RCPFrame => map(rcpframe, framebuffer_size)
-    )
-    pass2 = RenderObject(data2, shader2, PostprocessPrerender(), nothing)
-
-    pass2.postrenderfunction = () -> draw_fullscreen(pass2.vertexarray.id)
-
-    shader3 = LazyShader(
-        GLWindow.loadshader("fullscreen.vert"),
-        GLWindow.loadshader("copy.frag")
-    )
-
-    data3 = Dict{Symbol, Any}(
-        :color_texture => color
-    )
-
-    pass3 = RenderObject(data3, shader3, GLWindow.PostprocessPrerender(), nothing)
-
-    pass3.postrenderfunction = () -> GLWindow.draw_fullscreen(pass3.vertexarray.id)
-
-
-    (pass1, pass2, pass3)
-end
